# docs/openapi.yaml
openapi: 3.0.0
info:
  title: PKI Service API
  version: 1.0.0
  description: |
    Сервис электронной подписи документов с использованием ЭЦП РК (НУЦ РК).
    Поддерживает одиночное и пакетное подписание, интеграцию с корпоративным
    S3, BPM workflow и hash-based подписание через NCALayer.

    ## Аутентификация
    - `ApiKeyAuth` — `Authorization: Bearer <api_key>` (длинный токен без точек).
    - `BearerAuth` — JWT, выданный `/auth/login` или `/auth/register`.

    На `/api/v1/*` принимаются оба варианта (`DualAuth`). На `/auth/me` и
    `/auth/logout` — только JWT. На `/api/admin/*` — JWT + роль admin.

    ## Контуры
    - test: `http://localhost:8080`
    - prod: `https://pki-service.darch.pro`

    ## Примечания
    Некоторые request/response поля задокументированы на основе кода
    хендлеров; для нетривиальных потоков (sign/initiate, applications)
    точный JSON-контракт смотри в `internal/handler/*.go`.
  contact:
    email: admin@darch.pro
servers:
  - url: https://pki-service.darch.pro
    description: Production
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Логин
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Auth]
      security:
        - BearerAuth: []
      summary: Текущий пользователь
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  email: { type: string }
                  name: { type: string }
                  role: { type: string, enum: [user, admin] }
                  tenant_id: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      security:
        - BearerAuth: []
      summary: Логаут
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }

  # ───────── Verify (public HTML) ─────────
  /verify/{signature_id}:
    get:
      tags: [Verify]
      security: []
      summary: Публичная страница верификации подписи
      description: HTML-страница с данными подписи и QR-кодом.
      parameters:
        - name: signature_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: HTML-страница верификации
          content:
            text/html:
              schema: { type: string }
        '404':
          description: Подпись не найдена
          content:
            text/html:
              schema: { type: string }

  # ───────── Documents ─────────
  /api/v1/documents:
    post:
      tags: [Documents]
      summary: Создать запись документа по существующему S3-ключу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [s3_key]
              properties:
                s3_key: { type: string, example: tenant_uuid/path/file.pdf }
                title: { type: string, example: Договор № 12 }
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      status: { type: string, enum: [draft, pending, partially_signed, signed, rejected] }
                      created_at: { type: string, format: date-time }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/documents/{id}:
    get:
      tags: [Documents]
      summary: Получить документ + список подписей
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/documents/{id}/sign:
    post:
      tags: [Documents]
      summary: Синхронно подписать документ
      description: |
        Принимает CMS-подпись от NCALayer, верифицирует через NCANode,
        собирает новую версию PDF с QR-штампом и «Листом подписей»,
        обновляет статус документа.
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cms]
              properties:
                cms:
                  type: string
                  description: Base64 CMS (PKCS#7)
                role:
                  type: string
                  example: client
                  description: client / factor / director / ...
      responses:
        '200':
          description: Подписано
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      signature_id: { type: string, format: uuid }
                      signed_document_url: { type: string, example: tenant/doc/v2.pdf }
                      signature: { $ref: '#/components/schemas/Signature' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/documents/{id}/sign-async:
    post:
      tags: [Documents]
      summary: Поставить подпись в фон, вернуть 202
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cms]
              properties:
                cms: { type: string }
                role: { type: string }
      responses:
        '202':
          description: Принято, обработка в фоне
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      task_id: { type: string, format: uuid }
                      status: { type: string, example: processing }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/documents/{id}/sign-status:
    get:
      tags: [Documents]
      summary: Статус асинхронного подписания
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [pending, processing, signed, failed] }
                  signed_document_url: { type: string, nullable: true }
                  error: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/upload:
    post:
      tags: [Documents]
      summary: Загрузить файл документа в S3 и создать запись
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title: { type: string }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/documents/{id}/file:
    get:
      tags: [Documents]
      summary: Скачать текущую версию PDF
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────── Batch ─────────
  /api/v1/batch/upload:
    post:
      tags: [Batch]
      summary: Пакетная загрузка документов
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      documents:
                        type: array
                        items: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/batch/sign:
    post:
      tags: [Batch]
      summary: Пакетное подписание
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signatures]
              properties:
                signatures:
                  type: array
                  items:
                    type: object
                    required: [document_id, cms]
                    properties:
                      document_id: { type: string, format: uuid }
                      cms: { type: string }
                      role: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            document_id: { type: string, format: uuid }
                            status: { type: string, enum: [signed, failed] }
                            signature_id: { type: string, format: uuid, nullable: true }
                            error: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────── Session signing ─────────
  /api/v1/sign/initiate:
    post:
      tags: [SessionSigning]
      summary: Создать сессию подписания и поставить документы в очередь скачивания
      description: |
        Hash-flow (рекомендуется): клиент уже посчитал SHA-256 PDF и положил
        в `x-amz-meta-sha256` объекта S3 → передаёт `hash`, `s3_bucket`,
        `s3_key` в каждом документе → PKI не качает синхронно.
        Legacy-flow: только `source_url`; PKI скачает PDF и сам посчитает хэш.

        **Idempotency:** заголовок `Idempotency-Key` (произвольная строка ≤
        255 символов). Повтор с тем же ключом и Authorization-тенантом отдаёт
        кэшированный ответ + `X-Idempotent-Replay: true`. TTL 24h.

        **Mapping response.documents с request.documents:** используйте
        `client_ref` (если задан) или `client_index` — порядок гарантирован
        начиная с 2026-06; не полагайтесь на имя файла как ключ.
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/RequestID' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InitiateRequest' }
      responses:
        '200':
          description: Сессия создана, у части документов status=ready, у остальных fetch_failed
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            X-Idempotent-Replay:
              schema: { type: string }
              description: '`true` если ответ взят из idempotency-кэша'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InitiateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Дубликат документа (DUPLICATE_DOCUMENT_NAME или duplicate_client_ref)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '502':
          description: Все документы провалили fetch (FETCH_FAILED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/sign/complete:
    post:
      tags: [SessionSigning]
      summary: Завершить сессию — принять CMS, собрать PDF, залить в S3 клиента
      description: |
        CMS — base64 detached PKCS#7 (как возвращает NCALayer
        `kz.gov.pki.knca.basics.signHashes`). Поддерживается std-base64,
        raw-std-base64, url-safe-base64, raw-url-safe-base64. Внутри CMS
        проверяется messageDigest ↔ session.hash; mismatch → 422 HASH_MISMATCH.

        **Асинхронный постпроцессинг**: успешный ответ (`status: signed`)
        означает, что подпись криптографически подтверждена — это финальный
        юридический факт. QR-штамп на страницах, пересборка Листа подписей и
        выгрузка готового PDF клиенту при этом продолжаются в фоне. Ответ НЕ
        содержит `s3_key` и файл ещё недоступен для скачивания — используйте
        `GET /sign/status/{session_id}` и ждите `state: UPLOADED` для каждого
        документа, прежде чем показывать пользователю кнопку скачивания.
      parameters:
        - { $ref: '#/components/parameters/RequestID' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteRequest' }
      responses:
        '200':
          description: Завершено; per-document результат в `documents[]`
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompleteResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Сессия уже completed/failed/expired (SESSION_CLOSED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '410':
          description: Сессия истекла по TTL (SESSION_EXPIRED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422':
          description: |
            Бизнес-ошибка CMS. Возможные коды:
            `INVALID_CMS_BASE64`, `INVALID_CMS_STRUCTURE`, `HASH_MISMATCH`,
            `CERT_REVOKED`, `CERT_NOT_TRUSTED`, `CMS_INVALID`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/sign/status/{session_id}:
    get:
      tags: [SessionSigning]
      summary: Статус сессии
      description: |
        Возвращает агрегированный статус сессии + per-document состояние с
        нормализованным `state` enum для интеграций. Поле `status` сохранено
        для back-compat (raw внутренний статус).
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/RequestID' }
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionStatusResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sign/refresh-urls:
    patch:
      tags: [SessionSigning]
      summary: Перевыпустить presigned URL для документов сессии
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      documents:
                        type: array
                        items: { $ref: '#/components/schemas/SessionDocument' }

  # ───────── Applications ─────────
...
  /api/v1/applications/{external_id}/status:
    get:
      tags: [Applications]
      summary: Статус заявки
      parameters:
        - $ref: '#/components/parameters/ExternalID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Application' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API-ключ в заголовке `Authorization: Bearer <api_key>` (длинный токен без точек)."
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    AuthResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string }
            name: { type: string }
            role: { type: string, enum: [user, admin] }
            tenant_id: { type: string, format: uuid }
        token:
          type: string
          description: JWT, время жизни 24ч
...